+++ designsafe/apps/rapid/migrations/0001_initial.py
@@ -32,5 +32,4 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
-        migrations.RunPython(add_rapid_event_types)
     ]