+++ designsafe/static/scripts/data-depot/components/data-depot-listing/publication-advanced-search/publication-advanced-search-checkbox.component.js
@@ -1,0 +1,14 @@
+import publicationAdvancedSearchCheckboxTemplate from './publication-advanced-search-checkbox.template.html';
+
+
+export const publicationAdvancedSearchCheckbox = {
+    template: publicationAdvancedSearchCheckboxTemplate,
+    bindings: {
+        onSelect: '&',
+        selected: '<'
+    }
+
+}
+
+
+