+++ designsafe/static/scripts/data-depot/index.js
@@ -841,6 +841,30 @@ function config(
                 },
             },
         })
+        .state('publicDataLegacy', {
+            url: '/public-legacy/?typeFilters&query_string',
+            component: 'dataDepotPublicationsBrowser',
+            params: {
+                systemId: 'nees.public',
+                filePath: '',
+            },
+            resolve: {
+                apiParams: ()=> {
+                    return {
+                        fileMgr: 'public-legacy',
+                        baseUrl: '/api/public/files',
+                        searchState: 'publicDataLegacy',
+                    };
+                },
+                path: ($stateParams)=>{
+                    'ngInject';
+                    return $stateParams.filePath || '/';
+                },
+                auth: () => {
+                    return true;
+                },
+            },
+        })
         .state('neesPublished', {
             url: '/public/nees.public/{filePath:any}',
             component: 'neesPublished',
@@ -898,25 +922,42 @@ function config(
             },
             onExit: ($window) => {
                 'ngInject';
-                [
-                    'description', 'citation_title', 'citation_publication_date',
-                    'citation_doi', 'citation_abstract_html_url', 'DC.identifier', 
-                    'DC.creator', 'DC.title', 'DC.date'
-                ].forEach(
+                const projectLevelTags = [
+                    'description',
+                    'citation_title',
+                    'citation_publication_date',
+                    'citation_doi',
+                    'citation_abstract_html_url',
+                    'identifier',
+                    'DC.identifier',
+                    'DC.creator',
+                    'DC.title',
+                    'DC.date',
+                    'authors'
+                ];
+                projectLevelTags.forEach(
                     (name) => {
                         let el = $window.document.getElementsByName(name);
-                        el[0].content = '';
-                    }
-                );
-                ['citation_author', 'citation_author_institution',
-                    'citation_keywords'].forEach(
-                    (name) => {
-                        let els = $window.document.getElementsByName(name);
-                        while(els[0]){
-                            els[0].parentNode.removeChild(els[0]);
+                        if (el) {
+                            el[0].content = '';
+                            // citation_doi is also an entity tag
+                            while(el[1]) el[1].parentNode.removeChild(el[1]);
                         }
+                        
                     }
                 );
+                const entityTags = [
+                    'citation_author',
+                    'citation_author_institution',
+                    'citation_description',
+                    'citation_keywords',
+                ];
+                entityTags.forEach((name) => {
+                    let els = $window.document.getElementsByName(name);
+                    while (els[0]) {
+                        els[0].parentNode.removeChild(els[0]);
+                    }
+                });
             },
             resolve: {
                 listing: ($stateParams, DataBrowserService)=>{