+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # DesignSafe-CI Portal Release Notes
 
+## v4.9.9-20200703
+Fixes:
+- bug/DES-1698: Fixing typos in hybrid simulation templates
+- Fix Simulations Not Rendering
+
+## v4.9.9-20200702
+Enhancements:
+- Update Elasticsearch settings and reduce the amount of data retrieved from ES for simple listings.
+
+## v4.9.9
+Enhancements:
+- task/DES-1678: Curation UI/UX updates (1516, 1526, 1539)
+- task/DES-1632: Create meta tags for publications with multiple DOIs
+- task/DES-1670: Update MyData in WorkSpace Data Browser to match Data Depot
+- task/DES-1650: Add infinite scroll to Data Depot move/copy modals
+
+Fixes:
+- Removed call to custom script from rapid initial migration
+
+## v4.9.8-20200701
+Enhancements:
+- task/DES-1678: Curation UI/UX updates (1516, 1526, 1539)
+- task/DES-1632: Create meta tags for publications with multiple DOIs
+- task/DES-628: Advanced Search for Publications
+- task/DES-1670: Update MyData in WorkSpace Data Browser to match Data Depot
+- task/DES-1650: Add infinite scroll to Data Depot move/copy modals
+
+Fixes:
+- Removed call to custom script from rapid initial migration
+
+## v4.9.8
+Fixes:
+- Enable expansion of multiple tabs of the same type in Publications.
+- Fixes to the Authors window in the publication pipeline.
+
+## v4.9.7-20200602
+Fixes: 
+* Enable expansion of multiple tabs of the same type in Publications.
+* Fixes to the Authors window in the publication pipeline.
+
 ## v4.9.7
 Fixes:
 * Fix Recon Portal "Clear Filters" button.